# api/grpc/api/proto/swagger/v1/role/role.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "v1/role/role.proto",
    "version": "version not set"
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/api/v1/role": {
      "get": {
        "operationId": "Role_GetListRole",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/protoGetListRoleResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/runtimeError"
            }
          }
        },
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "field",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "Role"
        ]
      },
      "post": {
        "operationId": "Role_CreateRole",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/protoCreateRoleResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/runtimeError"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/protoCreateRoleRequest"
            }
          }
        ],
        "tags": [
          "Role"
        ]
      }
    },
    "/api/v1/role/{id}": {
      "get": {
        "operationId": "Role_GetRoleByID",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/protoGetRoleByIDResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/runtimeError"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "Role"
        ]
      },
      "delete": {
        "operationId": "Role_DeleteRole",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/protoDeleteRoleResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/runtimeError"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "Role"
        ]
      },
      "patch": {
        "operationId": "Role_UpdateRole",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/protoUpdateRoleResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/runtimeError"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/protoUpdateRoleRequest"
            }
          }
        ],
        "tags": [
          "Role"
        ]
      }
    }
  },
  "definitions": {
    "ResponseDataRole": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "createdAt": {
          "type": "string"
        },
        "createdBy": {
          "type": "string"
        },
        "updatedAt": {
          "type": "string"
        },
        "updatedBy": {
          "type": "string"
        }
      }
    },
    "protoCreateRoleRequest": {
      "type": "object",
      "properties": {
        "body": {
          "$ref": "#/definitions/protoCreateRoleRequestData"
        }
      }
    },
    "protoCreateRoleRequestData": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        }
      }
    },
    "protoCreateRoleResponse": {
      "type": "object",
      "properties": {
        "responseCode": {
          "type": "string"
        },
        "responseDesc": {
          "type": "string"
        },
        "responseData": {
          "$ref": "#/definitions/protoCreateRoleResponseResponseData"
        }
      }
    },
    "protoCreateRoleResponseResponseData": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        }
      }
    },
    "protoDeleteRoleResponse": {
      "type": "object",
      "properties": {
        "responseCode": {
          "type": "string"
        },
        "responseDesc": {
          "type": "string"
        },
        "responseData": {
          "$ref": "#/definitions/protoDeleteRoleResponseResponseData"
        }
      }
    },
    "protoDeleteRoleResponseResponseData": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        }
      }
    },
    "protoGetListRoleResponse": {
      "type": "object",
      "properties": {
        "responseCode": {
          "type": "string"
        },
        "responseDesc": {
          "type": "string"
        },
        "responseData": {
          "$ref": "#/definitions/protoGetListRoleResponseResponseData"
        }
      }
    },
    "protoGetListRoleResponseResponseData": {
      "type": "object",
      "properties": {
        "page": {
          "type": "integer",
          "format": "int32"
        },
        "limit": {
          "type": "integer",
          "format": "int32"
        },
        "total": {
          "type": "integer",
          "format": "int32"
        },
        "totalPage": {
          "type": "integer",
          "format": "int32"
        },
        "roles": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ResponseDataRole"
          }
        }
      }
    },
    "protoGetRoleByIDResponse": {
      "type": "object",
      "properties": {
        "responseCode": {
          "type": "string"
        },
        "responseDesc": {
          "type": "string"
        },
        "responseData": {
          "$ref": "#/definitions/protoGetRoleByIDResponseResponseData"
        }
      }
    },
    "protoGetRoleByIDResponseResponseData": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        }
      }
    },
    "protoUpdateRoleRequest": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "body": {
          "$ref": "#/definitions/protoUpdateRoleRequestData"
        }
      }
    },
    "protoUpdateRoleRequestData": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        }
      }
    },
    "protoUpdateRoleResponse": {
      "type": "object",
      "properties": {
        "responseCode": {
          "type": "string"
        },
        "responseDesc": {
          "type": "string"
        },
        "responseData": {
          "$ref": "#/definitions/protoUpdateRoleResponseResponseData"
        }
      }
    },
    "protoUpdateRoleResponseResponseData": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "type_url": {
          "type": "string"
        },
        "value": {
          "type": "string",
          "format": "byte"
        }
      }
    },
    "runtimeError": {
      "type": "object",
      "properties": {
        "error": {
          "type": "string"
        },
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}